refactor: Split them into multiple files because performance is not a problem anymore
